Fixed for 64bit compilation
git-svn-id: https://svn.aros.org/svn/aros/trunk@26464 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Feelin/Sources/Group/object.c
@@ -9,7 +9,7 @@
 
 */
 
-void Group_DoA(FClass *Class,FObject Obj,ULONG Method,APTR Msg)
+void Group_DoA(FClass *Class,FObject Obj,uint32 Method,APTR Msg)
 {
    struct LocalObjectData *LOD = F_LOD(Class,Obj);
